Bug fixing

+++ Week04/functions_tarik_bozgan.py
@@ -1,3 +1,4 @@
+# x parametresine =0 ekledik, testin istediği bu.
 custom_power = lambda x=0, /, e=1: x ** e
 
 def custom_equation(x: int = 0, y: int = 0, /, a: int = 1, b: int = 1, *, c: int = 1) -> float:
@@ -15,7 +16,7 @@ def custom_equation(x: int = 0, y: int = 0, /, a: int = 1, b: int = 1, *, c: int
     if not isinstance(b, int): raise TypeError("b must be int")
     if not isinstance(c, int): raise TypeError("c must be int")
     
-    return (x ** a + y ** b) / c
+    return float((x ** a + y ** b) / c)
 
 _count = 0
 def fn_w_counter() -> (int, dict[str, int]):
